chore: split eth_calls collector into submodules
